Fixed Health reward and Score reward
Score reward needed testing and fixing, after last merge.
Health reward needed testing and fixing, after commit
https://github.com/Mtihc/TreasureChest/commit/769fcb7046ad9a88546294c883dc4bb41ed83f3f

Score reward added! Fixes
https://github.com/Mtihc/TreasureChest/issues/48

+++ src/main/java/com/mtihc/minecraft/treasurechest/v8/rewardfactory/rewards/HealthReward.java
@@ -17,7 +17,7 @@ public class HealthReward implements IReward {
 		this.info = info;
 	}
 	
-	public HealthReward(int health) {
+	public HealthReward(double health) {
 		Map<String, Object> data = new HashMap<String, Object>();
 		data.put("health", health);
 		this.info = new RewardInfo("health", data);
@@ -29,7 +29,7 @@ public class HealthReward implements IReward {
 	}
 	
 	public double getHealth() {
-		return (Double) info.getData("health");
+		return (Integer) info.getData("health");
 	}
 	
 	public void setHealth(double value) {
